@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <RootNamespace>ReInject.PostInjectors.BackgroundWorker</RootNamespace>
     <PackageId>Aeolin.reInject.PostInjectors.BackgroundWorker</PackageId>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
@@ -12,17 +12,20 @@
     <RepositoryUrl>https://github.com/re-develop/reInject</RepositoryUrl>
     <RepositoryType>git</RepositoryType>
     <PackageTags>di dependency injection background worker scheduler</PackageTags>
-    <AssemblyVersion>1.5.4.2</AssemblyVersion>
-    <FileVersion>1.5.4.2</FileVersion>
+    <AssemblyVersion>2.0.0</AssemblyVersion>
+    <FileVersion>2.0.0</FileVersion>
     <Copyright>Aeolin Ferjünnoz</Copyright>
     <Description>An extension for reInject adding post injectable background workers</Description>
-    <Version>1.5.4.2</Version>
+    <Version>2.0.0</Version>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Aeolin.reInject" Version="1.5.3" />
     <PackageReference Include="Cronos" Version="0.7.1" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\reInject\ReInject.csproj" />
   </ItemGroup>
 
 </Project>
