@@ -1,10 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <PackageId>Aeolin.$(AssemblyName)</PackageId>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <Version>1.5.3.1</Version>
+    <Version>2.0.0</Version>
     <Company>reDevelop</Company>
     <Authors>Aeolin Ferjünnoz</Authors>
     <RepositoryUrl>https://github.com/re-develop/reInject</RepositoryUrl>
@@ -13,7 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Aeolin.reInject" Version="1.5.3" />
+    <ProjectReference Include="..\reInject\ReInject.csproj" />
   </ItemGroup>
 
 </Project>
