@@ -1,26 +1,29 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <RepositoryUrl>https://github.com/re-develop/reInject</RepositoryUrl>
     <RepositoryType>git</RepositoryType>
     <PackageTags>di dependency injection dependencyinjection reInject</PackageTags>
     <NeutralLanguage>en</NeutralLanguage>
-    <AssemblyVersion>1.5.3</AssemblyVersion>
-    <FileVersion>1.5.3</FileVersion>
+    <AssemblyVersion>2.0.0</AssemblyVersion>
+    <FileVersion>2.0.0</FileVersion>
     <Company>reDevelop</Company>
     <Authors>Aeolin Ferjünnoz</Authors>
     <Product>reInject.Scopes</Product>
     <Description>IServiceScope and IServiceScopeFactory Implementation for reinject</Description>
     <Copyright>Aeolin Ferjünnoz</Copyright>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <Version>1.5.3</Version>
+    <Version>2.0.0</Version>
     <PackageId>Aeolin.reInject.Scopes</PackageId>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Aeolin.reInject" Version="1.5.3" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\reInject\ReInject.csproj" />
   </ItemGroup>
 
 </Project>
