@@ -12,38 +12,28 @@ namespace ReInject.Implementation.DependencyTypes
   /// <summary>
   /// internal abstract base class to implement IDependecyType
   /// </summary>
-  internal abstract class DependencyBase : IDependencyType
+  internal abstract class DependencyBase<T> : IDependency
   {
-    public DependencyBase( IDependencyContainer container, DependencyStrategy strategy, Type type, Type interfaceType = null, string name = null )
+    public DependencyBase( IDependencyContainer container, Func<T> factory, Type type, string name = null )
     {
-      this.Strategy = strategy;
       this.Type = type;
-      this.InterfaceType = interfaceType;
       this._container = container;
       this.Name = name;
+      _factory = factory;
     }
 
 
     public string Name { get; init; }
-
-
+    private Func<T> _factory;
     private IDependencyContainer _container;
-    public DependencyStrategy Strategy { get; private set; }
-
     public Type Type { get; private set; }
-
-    public Type InterfaceType { get; private set; }
-
-    public abstract object Instance { get; }
-
+    public virtual object Instance => createInstance();
     public abstract void Clear();
-
-
-
+    public virtual bool IsSingleton => false;
 
     protected object createInstance()
     {
-      return TypeInjectionMetadataCache.GetMetadataCache(Type).CreateInstance(_container);
+      return _factory != null ? _factory.Invoke() : TypeInjectionMetadataCache.GetMetadataCache<T>().CreateInstance(_container);
     }
 
   }
