@@ -34,75 +34,13 @@ namespace ReInjectTests
     }
   }
 
-  public class EventSource
-  {
-    public delegate int TestDelegate(int num, string test, object obj1, object obj2, TestData obj3);
-    public event TestDelegate TestEvent;
-    public event Action<string, int> SimpleEvent;
-
-    public struct TestData
-    {
-      public int Field1;
-      public string Field2;
-    }
-
-    public int CallEvent(int num)
-    {
-      SimpleEvent?.Invoke("test", num);
-      return TestEvent?.Invoke(num, "lol", null, null, new TestData() { Field1 = 13, Field2 = "test" }) ?? -1;
-    }
-  }
-
-  public class EventTarget
-  {
-    public int LastEventValue { get; private set; }
-    public string LastSimpleEventValue { get; private set; }
-    public int CountSimpleEventCalled { get; private set; } = 0;
-
-    [InjectEvent("test")]
-    public int HandleEvent(int num, string test, object obj1, object obj2, TestData obj3)
-    {
-      LastEventValue = num;
-      return num * 2;
-    }
-
-    [InjectEvent("simple")]
-    [InjectEvent("simple2")]
-    public void HandleSimple(string str, int num)
-    {
-      LastSimpleEventValue = str + num;
-      CountSimpleEventCalled++;
-    }
-
-  }
+ 
 
 
   public class ReInjectIntegrationTests
   {
 
-    [Fact]
-    public void ReInject_TestEventInjection_EventCallbackWorks()
-    {
-      // Arrange
-      int num = 1337;
-      var source = new EventSource();
-      var target = new EventTarget();
-
-      var proxy = new EventProxy(source, nameof(source.TestEvent), "test");
-      var proxy2 = new EventProxy(source, nameof(source.SimpleEvent), "simple");
-      var proxyTarget = new EventProxyTarget(target, ReflectionHelper.GetMember<MethodInfo>(target.GetType(), "HandleEvent"));
-      var proxyTarget2 = new EventProxyTarget(target, ReflectionHelper.GetMember<MethodInfo>(target.GetType(), "HandleSimple"));
-      proxy.AddTarget(proxyTarget);
-      proxy2.AddTarget(proxyTarget2);
-
-      // Act
-      var res = source.CallEvent(num);
-
-      // Assert
-      Assert.Equal(num, target.LastEventValue);
-      Assert.Equal(num * 2, res);
-      Assert.Equal($"test{num}", target.LastSimpleEventValue);
-    }
+  
 
     class ReadOnlyPropertyClass
     {
@@ -116,7 +54,7 @@ namespace ReInjectTests
       // Arrange
       var container = Injector.GetContainer(Guid.NewGuid().ToString());
       var num = 1337;
-      container.Register<int>(DependencyStrategy.AtomicInstance, true, num, "num");
+      container.AddSingleton(num, true, "num");
 
       // Act
       var instance = container.GetInstance<ReadOnlyPropertyClass>();
@@ -127,66 +65,13 @@ namespace ReInjectTests
     }
 
 
-    [Fact]
-    public void ReInject_TestEventInjection_DisablingEventsWorks()
-    {
-      // Arrange
-      var container = Injector.GetContainer(Guid.NewGuid().ToString());
-      int num = 1337;
-      var source = new EventSource();
-      container.AddEventInjector(setup =>
-      {
-        setup.RegisterEventSource(source, "TestEvent", "test");
-      });
-
-      // Act
-      var target = container.GetInstance<EventTarget>();
-      var res0 = source.CallEvent(num);
-      container.SetEventTargetEnabled(target, false, "test");
-      var res1 = source.CallEvent(num);
-      container.SetEventTargetEnabled(target, true);
-      var res2 = source.CallEvent(num);
-
-
-
-      // Assert
-      Assert.Equal(num, target.LastEventValue);
-      Assert.Equal(num * 2, res0);
-      Assert.Equal(default(int), res1);
-      Assert.Equal(num * 2, res2);
-    }
-
-    [Fact]
-    public void ReInject_TestEventInjection_AttributeBasedInjectionWorks()
-    {
-      // Arrange
-      var container = Injector.GetContainer(Guid.NewGuid().ToString());
-      int num = 1337;
-      var source = new EventSource();
-      container.AddEventInjector(setup =>
-      {
-        setup.RegisterEventSource(source, "TestEvent", "test")
-        .RegisterEventSource(source, "SimpleEvent", "simple2")
-        .RegisterEventSource(source, "SimpleEvent", "simple");
-      });
-
-      // Act
-      var target = container.GetInstance<EventTarget>();
-      var res = source.CallEvent(num);
-
-
-      // Assert
-      Assert.Equal($"test{num}", target.LastSimpleEventValue);
-      Assert.Equal(2, target.CountSimpleEventCalled);
-      Assert.Equal(num, target.LastEventValue);
-      Assert.Equal(num * 2, res);
-    }
+   
 
     [Fact]
     public void ReInject_HandelsAbstract()
     {
       var container = Injector.GetContainer(Guid.NewGuid().ToString());
-      container.Register<IHelloService>(DependencyStrategy.AtomicInstance, true, new HelloService());
+      container.AddSingleton<IHelloService, HelloService>(true);
       container.GetInstance<StupidCtor>();
     }
 
@@ -194,7 +79,8 @@ namespace ReInjectTests
     public void ReInject_CreatesIntance_IfTypesAreGiven()
     {
       var container = Injector.GetContainer(Guid.NewGuid().ToString()); // get independent container
-      container.Register<IHelloService, HelloService>().Register<GoodByeService>(DependencyStrategy.NewInstance);
+      container.AddTransient<IHelloService, HelloService>()
+        .AddTransient<GoodByeService>();
       var inst = container.GetInstance<Greeter>();
       var msg = inst.Greet();
 
@@ -206,9 +92,9 @@ namespace ReInjectTests
     public void ReInject_UsesNamedParameter_IfTypesAndNameAreGiven()
     {
       var container = Injector.GetContainer(Guid.NewGuid().ToString());// get independent container
-      container.Register<IHelloService, HelloService>().Register<GoodByeService>(DependencyStrategy.NewInstance);
+      container.AddTransient<IHelloService, HelloService>().AddTransient<GoodByeService>();
       var inst = container.GetInstance<Greeter>();
-      container.Register<IHelloService, HelloService2>(name: "hello");
+      container.AddTransient<IHelloService, HelloService2>(name: "hello");
       container.PostInject(inst);
       var msg = inst.Greet();
 
@@ -220,13 +106,11 @@ namespace ReInjectTests
     {
       // Setup
       var container = Injector.GetContainer();
-      var dep = new NewInstanceDependency(container, typeof(HelloService), typeof(IHelloService));
+      var dep = new TransientDependency<HelloService>(container, null, typeof(IHelloService));
 
       // Act
       var inst1 = dep.Instance;
       var inst2 = dep.Instance;
-
-
 
       // Assert
       Assert.NotEqual(inst1, inst2);
@@ -236,7 +120,11 @@ namespace ReInjectTests
     public void InjectAttributeType_OverridesAutomaticDetectedType_IfAttributeFieldTypeIsSet()
     {
       var container = Injector.GetContainer(Guid.NewGuid().ToString());// get independent container
-      container.Register<IHelloService, HelloService>().Register<HelloService2>().Register<GoodByeService>(DependencyStrategy.NewInstance);
+      container
+        .AddTransient<IHelloService, HelloService>()
+        .AddTransient<HelloService2>()
+        .AddTransient<GoodByeService>();
+
       var inst = container.GetInstance<Greeter2>();
       var msg = inst.Greet();
 
